update and clean up
