+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/20 10:13:34 by spyun         #+#    #+#                 */
-/*   Updated: 2025/01/20 15:46:57 by spyun         ########   odam.nl         */
+/*   Updated: 2025/01/20 15:48:36 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,27 @@
 # include <readline/readline.h>
 
 extern int	g_exit_status;
+
+typedef enum e_token_type
+{
+	TOKEN_WORD,
+	TOKEN_PIPE, // |
+	TOKEN_REDIR_IN, // <
+	TOKEN_REDIR_OUT, // >
+	TOKEN_HEREDOC, // <<
+	TOKEN_APPEND, // >>
+	TOKEN_AND, // &&
+	TOKEN_OR, // ||
+	TOKEN_LPAREN, // (
+	TOKEN_RPAREN
+}	t_token_type;
+
+typedef struct s_token
+{
+	char			*content;
+	t_token_type	type;
+	struct s_token	*next;
+}	t_token;
 
 typedef struct s_cmd
 {
